Adds edit listing functionality
User can edit a listing by clicking the edit button next to the listing
on /listings. A get request is made to api for the listing ID and the
form is prepopulated with the data for the selected listing. User can
edit the game for trade or wanted game then submit. This returns user
to listings where they will see the updated listing.

This is a triumph.